Fix MSVC compiler warnings
* reg. int vs. size_t (C4267)
* reg. unused function parameters (C4100), introduced by #16062

Todo: a warning reg. conversion of float to _Rx (C4244), introduced by #16062

+++ src/engraving/compat/midi/midirender.cpp
@@ -613,7 +613,7 @@ static void collectGraceBeforeChordEvents(Chord* chord, EventMap* events, const 
     size_t acciacaturaGraceSize = graceNotesBeforeBar.size();
     if (acciacaturaGraceSize > 0) {
         graceTickSum = graceNotesBeforeBar[0]->ticks().ticks();
-        graceTickOffset = graceTickSum / acciacaturaGraceSize;
+        graceTickOffset = graceTickSum / static_cast<int>(acciacaturaGraceSize);
     }
 
     if (!graceNotesMerged(chord)) {
